+++ Changelog.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.  
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+# [1.4.2] - 2022-02-22
+## Added
+- Added reviewdog.yml in .github/workflow
+## Changed
+- Changed the CLI package configuration for cobra.
 # [1.4.1] - 2022-01-31
 ## Added
 - Added GOARCH to the generated Makefile